reciprocal function test for numeric values

+++ ivy_tests/test_ivy/test_frontends/test_tensorflow/test_math.py
@@ -449,7 +449,7 @@ def test_tensorflow_log_sigmoid(
 @handle_frontend_test(
     fn_tree="tensorflow.math.reciprocal",
     dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
+        available_dtypes=helpers.get_dtypes("numeric"),
     ),
     test_with_out=st.just(False),
 )
